feat: 🚀 Speed up sprite generation 40x
The previous convert-svg-to-png library used `puppeteer.screenshot` to convert an SVG to a PNG.
This method takes >200ms for each sprite on a MacOS. This replaces this with svg-to-img which
converts the SVG using canvas in puppeteer, which is an order of magnitude faster.

Also was able to simplify the async code by using promises and async/await, and take advantage
of the internal queuing in svg-to-img.