feat: Add scripts to extract messages and build translations (#14)
mapeo-settings extract-messages -o <file>
Will output a messages file of the format:
{
  [messageId: string]: { description: string, message: string }
}
that can be used as input for CrowdIn

mapeo-settings build will now look for valid message files in the folder
`messages` and build them into `translations.json` included in the
settings file. `translations.json` is the format expected by iD Editor,
for Mapeo it needs to be flattened to the format
`{ [messageId: string]: message }` so that react-intl can understand it.